update folder structure
