+++ apps/site/src/pages/account/api-keys-panel/api-keys-list/shared/format-date.ts
@@ -1,6 +1,6 @@
 import { format, formatDistanceToNowStrict, isSameDay } from "date-fns";
 
-export const formatDateRelativeAndExact = (date?: Date | string | null) => {
+export const formatDate = (date?: Date | string | null) => {
   if (!date) {
     return { relative: "Never", exact: "" };
   }
